pyproject.toml
[tool.poetry]
name = "mr-dale"
version = "0.1.0"
description = "discord bot that will always come to the rescue when working together"
license = "MIT"
authors = ["evgeniy tulyakov <e.v.tulyakov1998@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
python-dotenv = "^0.17.1"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.2"
flake8-import-order = "^0.18.1"
flake8-quotes = "^3.2.0"
flake8-variables-names = "^0.0.4"


[tool.poetry.scripts]
mr_dale = "mr_dale.server:start"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
